requirements.txt
pandas==1.5.3
pip==23.0.1
selenium==4.8.3
webdriver-manager==3.8.5
Flask
mysql-connector-python
sqlalchemy